requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
coverage==6.5.0
Django==4.1.4
django-environ==0.9.0
django-fsm==2.8.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
numpy==1.24.0
psycopg2-binary==2.9.5
PuLP==2.7.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
